refactor: fix controller resposibility
divide rest API controller, move Rooms object from controller layer to service layer
