+++ src/main/java/chess/service/RoomService.java
@@ -1,6 +1,8 @@
 package chess.service;
 
 import chess.dao.RoomDAO;
+import chess.domain.ChessGame;
+import chess.domain.Rooms;
 import chess.dto.RoomDTO;
 import org.springframework.stereotype.Service;
 
@@ -10,9 +12,11 @@ import java.util.List;
 public final class RoomService {
 
     private final RoomDAO roomDAO;
+    private final Rooms rooms;
 
-    public RoomService(final RoomDAO roomDAO) {
+    public RoomService(final RoomDAO roomDAO, final Rooms rooms) {
         this.roomDAO = roomDAO;
+        this.rooms = rooms;
     }
 
     public List<RoomDTO> allRooms() {
@@ -29,5 +33,13 @@ public final class RoomService {
 
     public List<String> allRoomsId() {
         return roomDAO.allRoomIds();
+    }
+
+    public void addRoom(final String roomId, final ChessGame chessGame) {
+        rooms.addRoom(roomId, chessGame);
+    }
+
+    public ChessGame loadGameByRoomId(final String roomId) {
+        return rooms.loadGameByRoomId(roomId);
     }
 }
