+++ src/main/resources/templates/chess.hbs
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Woowa Kingdom Chess</title>
-    <link rel="stylesheet" type="text/css" href="css/chess.css">
+    <link rel="stylesheet" type="text/css" href="/css/chess.css">
 </head>
 <body>
 <div id="total">
@@ -135,7 +135,7 @@
     </div>
 </div>
 <script src="http://code.jquery.com/jquery-latest.js"></script>
-<script type="text/javascript" src="js/chess.js"></script>
+<script type="text/javascript" src="/js/chess.js"></script>
 <script type="text/javascript">
     {{#pieces}}
     document.getElementById("{{position}}")
